test: 진행 중
- 이미지 업로드 및 고민 등록
- 고민 답변
- 검색 조회 페이지네이션

+++ gogoApp/app/src/main/java/com/sixthank/gogo/src/main/home/models/HomeResponse.java
@@ -47,9 +47,9 @@ public class HomeResponse {
         @SerializedName("endDateTime")
         private String mEndDateTime;
         @SerializedName("id")
-        private Long mId;
+        private int mId;
         @SerializedName("memberId")
-        private Long mMemberId;
+        private int mMemberId;
         @SerializedName("pictureUrl")
         private String mPictureUrl;
         @SerializedName("startDateTime")
@@ -73,19 +73,19 @@ public class HomeResponse {
             mEndDateTime = endDateTime;
         }
 
-        public Long getId() {
+        public int getId() {
             return mId;
         }
 
-        public void setId(Long id) {
+        public void setId(int id) {
             mId = id;
         }
 
-        public Long getMemberId() {
+        public int getMemberId() {
             return mMemberId;
         }
 
-        public void setMemberId(Long memberId) {
+        public void setMemberId(int memberId) {
             mMemberId = memberId;
         }
